Moving store var to types
